Update and rename fkbpro-init.js to fkb-template.js
