+++ javascripts/discourse/api-initializers/fkb-template.js
@@ -1,0 +1,22 @@
+import { withPluginApi } from "discourse/lib/plugin-api";
+import { findRawTemplate } from "discourse-common/lib/raw-templates";
+import { htmlSafe } from "@ember/template";
+
+export default {
+  name: "fkbpro",
+  initialize() {
+    withPluginApi("0.8.7", (api) => {
+      // Use same template on Desktop and MobileView
+      api.modifyClass("component:topic-list-item", {
+        pluginId: "fkb-template",
+
+        renderTopicListItem() {
+          const template = findRawTemplate("list/custom-topic-list-item");
+          if (template) {
+            this.set("topicListItemContents", htmlSafe(template(this)));
+          }
+        },
+      });
+    });
+  },
+};
